Refactored and optimised code with useMemo and useCallback hooks
