+++ src/App.js
@@ -1,9 +1,10 @@
-import { useState, useEffect, useContext } from 'react';
+import { useContext, useMemo } from 'react';
 
 import { Routes, Route } from 'react-router-dom';
 
 import ThemeContext from './store/ThemeProvider';
-import CountryContext from './store/CountryProvider';
+
+import { useFetchCountries } from './hooks/useFetchCountries';
 
 import Home from './pages/Home';
 import CountryPage from './pages/CountryPage';
@@ -11,72 +12,13 @@ import Header from './components/Header/Header';
 import Loading from './UI/Loading';
 import Error from './UI/Error';
 
-const transformData = (data) => {
-	const countries = data.map((country) => {
-		return {
-			alpha3Code: country.alpha3Code,
-			id: country.name,
-			name: country.name,
-			nativeName: country.nativeName,
-			population: country.population.toLocaleString(),
-			region: country.region,
-			subregion: country.subregion,
-			capital: !country?.capital ? 'No capital' : country.capital,
-			topLevelDomain:
-				country.topLevelDomain[0] === ''
-					? 'No domain'
-					: country.topLevelDomain[0],
-			currencies: !country?.currencies
-				? 'No currency'
-				: country.currencies.map((currency) => currency.name),
-			languages: country.languages.map((lang) => lang.name),
-			borders: country.borders,
-			flags: country.flags.svg,
-			slug: country.alpha3Code.toLowerCase(),
-			shortName:
-				country.name.length <= 10
-					? country.name
-					: `${country.name.slice(0, 10)}...`,
-		};
-	});
-	return countries;
-};
-
 const App = () => {
-	const [loading, setLoading] = useState(false);
-	const [error, setError] = useState(false);
 	const { darkTheme } = useContext(ThemeContext);
-	const { createCountriesArray } = useContext(CountryContext);
 
-	const fetchCountries = async () => {
-		setLoading(true);
-		setError(null);
+	const { error, loading } = useFetchCountries();
 
-		const countriesArr = JSON.parse(window.localStorage.getItem('countries'));
-
-		if (countriesArr) createCountriesArray(countriesArr);
-
-		if (!countriesArr) {
-			try {
-				const res = await fetch('https://restcountries.com/v2/all');
-				const data = await res.json();
-
-				const transformedCountries = transformData(data);
-
-				createCountriesArray(transformedCountries);
-
-				localStorage.setItem('countries', JSON.stringify(transformedCountries));
-			} catch (err) {
-				setError("Couldn't find countries, please reload the page!");
-			}
-		}
-
-		setLoading(false);
-	};
-
-	useEffect(() => {
-		fetchCountries();
-	}, []);
+	// memoized Header so it doesn't re-render when new page is rendered
+	const header = useMemo(() => <Header />, []);
 
 	const containerClass = `container container--home ${
 		!darkTheme && 'light-mode'
@@ -89,7 +31,7 @@ const App = () => {
 
 	return (
 		<div className={containerClass}>
-			<Header />
+			{header}
 
 			<Routes>
 				<Route path="/" element={content} />
