+++ src/store/CountryProvider.js
@@ -1,4 +1,4 @@
-import { createContext, useState } from 'react';
+import { createContext, useState, useMemo, useCallback } from 'react';
 
 const CountryContext = createContext();
 
@@ -9,52 +9,73 @@ export const CountryProvider = ({ children }) => {
 	const [filter, setFilter] = useState(false);
 
 	// creates an array containing all country objects
-	const createCountriesArray = (data) => {
-		setCountriesArray(data);
-	};
+	const createCountriesArray = useCallback(
+		(data) => {
+			setCountriesArray(data);
+		},
+		[setCountriesArray]
+	);
 
 	// gets user back to the homepage
-	const goBackHandler = () => {
+	const goBackHandler = useCallback(() => {
 		setFilteredCountriesArray(countriesArray);
-	};
+	}, [setFilteredCountriesArray, countriesArray]);
 
 	// creates new array of country objects filtered by region
-	const filterByRegionHandler = (region) => {
-		if (region === 'America') region = 'Americas';
+	const filterByRegionHandler = useCallback(
+		(region) => {
+			if (region === 'America') region = 'Americas';
 
-		const filteredCountries = countriesArray.filter(
-			(country) => country.region === region
-		);
+			const filteredCountries = countriesArray.filter(
+				(country) => country.region === region
+			);
 
-		setFilteredCountriesArray(filteredCountries);
-		setFilter(true);
-	};
+			setFilteredCountriesArray(filteredCountries);
+			setFilter(true);
+		},
+		[countriesArray, setFilter]
+	);
 
 	// creates new array of country objects filtered by search result
-	const searchCountryHandler = (search) => {
-		if (search === '') setFilteredCountriesArray(countriesArray);
+	const searchCountryHandler = useCallback(
+		(search) => {
+			if (search === '') setFilteredCountriesArray(countriesArray);
 
-		const searchedCountries = countriesArray.filter((country) =>
-			country.name.toLowerCase().includes(search.toLowerCase())
-		);
+			const searchedCountries = countriesArray.filter((country) =>
+				country.name.toLowerCase().includes(search.toLowerCase())
+			);
 
-		setFilteredCountriesArray(searchedCountries);
-		setFilter(true);
-	};
+			setFilteredCountriesArray(searchedCountries);
+			setFilter(true);
+		},
+		[setFilteredCountriesArray, countriesArray, setFilter]
+	);
+
+	const memoizedValues = useMemo(
+		() => ({
+			countriesArray,
+			createCountriesArray,
+			goBackHandler,
+			filterByRegionHandler,
+			filteredCountriesArray,
+			filter,
+			setFilter,
+			searchCountryHandler,
+		}),
+		[
+			countriesArray,
+			createCountriesArray,
+			goBackHandler,
+			filterByRegionHandler,
+			filteredCountriesArray,
+			filter,
+			setFilter,
+			searchCountryHandler,
+		]
+	);
 
 	return (
-		<CountryContext.Provider
-			value={{
-				countriesArray,
-				createCountriesArray,
-				goBackHandler,
-				filterByRegionHandler,
-				filteredCountriesArray,
-				filter,
-				setFilter,
-				searchCountryHandler,
-			}}
-		>
+		<CountryContext.Provider value={memoizedValues}>
 			{children}
 		</CountryContext.Provider>
 	);
