+++ src/store/ThemeProvider.js
@@ -1,16 +1,21 @@
-import { createContext, useState } from 'react';
+import { createContext, useState, useMemo, useCallback } from 'react';
 
 const ThemeContext = createContext();
 
 export const ThemeProvider = ({ children }) => {
 	const [darkTheme, setDarkTheme] = useState(true);
 
-	const changeTheme = () => {
+	const changeTheme = useCallback(() => {
 		setDarkTheme((prevTheme) => !prevTheme);
-	};
+	}, [setDarkTheme]);
+
+	const memoizedValues = useMemo(
+		() => ({ darkTheme, changeTheme }),
+		[darkTheme, changeTheme]
+	);
 
 	return (
-		<ThemeContext.Provider value={{ darkTheme, changeTheme }}>
+		<ThemeContext.Provider value={memoizedValues}>
 			{children}
 		</ThemeContext.Provider>
 	);
